+++ src/components/Activity/components/components/tests/VizHighestRated.test.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { VizHighestRated } from "..";
 import type { MediaDiaryWithId } from "../../../../../types/typesMedia";
 import { Image } from "@chakra-ui/react";
-import { mountWithContext } from "@/utils/test-utils";
+import { mountWithContext } from "../../../../../utils/test-utils";
 
 const list: MediaDiaryWithId[] = [
   {
