+++ src/components/Content/Content.tsx
@@ -1,39 +1,21 @@
-import {
-  Box,
-  Grid,
-  Modal,
-  ModalBody,
-  ModalCloseButton,
-  ModalContent,
-  ModalHeader,
-  ModalOverlay,
-  useDisclosure,
-} from "@chakra-ui/react";
+import { Box, Grid, useDisclosure } from "@chakra-ui/react";
 import Head from "next/head";
 import { Router } from "next/router";
 import type { PropsWithChildren } from "react";
-import React, { useEffect, useRef } from "react";
-import { useMDDispatch, useMDState } from "../../config/store";
+import React, { useEffect } from "react";
+import { Day, Edit, Log, Search, Selected } from "..";
+import { useMDState } from "../../config/store";
 import useIsBreakpoint from "../../utils/useIsBreakpoint";
 import Layout from "../layouts/Layout";
-import MdLogo from "../md/MdLogo";
-import { Day, Edit, Log, Selected } from "..";
 import MdLoader from "../md/MdLoader";
-import {
-  ContentDrawer,
-  ContentSearch,
-  ContentSidebar,
-  ContentToolbar,
-} from "./components";
+import { ContentDrawer, ContentSidebar, ContentToolbar } from "./components";
 
 export default function Content({
   children,
   title = "MediaDiary",
 }: PropsWithChildren<unknown> & { title: string }): JSX.Element {
   const { view } = useMDState();
-  const dispatch = useMDDispatch();
   const isMd = useIsBreakpoint("md");
-  const refInput = useRef<HTMLInputElement>(null);
   const { isOpen, onClose, onOpen } = useDisclosure();
   const [loading, setLoading] = React.useState(false);
   // route transition - in order to prevent bad UX because of SSR props
@@ -84,25 +66,7 @@ export default function Content({
             {view === "log" && <Log />}
             {view === "edit" && <Edit />}
           </ContentDrawer>
-          <Modal
-            isOpen={view === "search"}
-            onClose={() => dispatch({ type: "close" })}
-            scrollBehavior="inside"
-            size={isMd ? "xl" : "sm"}
-            initialFocusRef={refInput}
-          >
-            <ModalOverlay sx={{ zIndex: 2 }}>
-              <ModalContent maxHeight="50vh" my={{ base: 2, sm: "3.75rem" }}>
-                <ModalCloseButton />
-                <ModalHeader>
-                  <MdLogo title="Search" />
-                </ModalHeader>
-                <ModalBody pt={0} pb={6}>
-                  {view === "search" && <ContentSearch refInput={refInput} />}
-                </ModalBody>
-              </ModalContent>
-            </ModalOverlay>
-          </Modal>
+          <Search />
         </>
       )}
     </Layout>
