--- src/components/Content/components/ContentSearch/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./ContentSearch";
