+++ src/components/Content/components/index.tsx
@@ -1,5 +1,4 @@
 export { default as ContentDrawer } from "./ContentDrawer";
 export { default as ContentFilters } from "./ContentFilters";
-export { default as ContentSearch } from "./ContentSearch";
 export { default as ContentSidebar } from "./ContentSidebar";
 export { default as ContentToolbar } from "./ContentToolbar";
