+++ src/components/Day/components/DayContent/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DayContent";
