+++ src/components/Search/components/SearchContent/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SearchContent";
