+++ src/components/Search/components/index.ts
@@ -1,0 +1,1 @@
+export { default as SearchContent } from "./SearchContent";
