+++ src/config/index.ts
@@ -1,4 +1,5 @@
 export * from "./contants";
 export * from "./store";
 export * from "./storeLog";
-export * from "./useDataFetch";
+export { default as useDataFetch } from "./useDataFetch";
+export type { DataFetch, DataFetchMDb, DataFetchSpotify } from "./useDataFetch";
