+++ src/utils/index.ts
@@ -4,4 +4,3 @@ export * from "./helperSpotify";
 export * from "./index";
 export * from "./useDebounce";
 export * from "./useIsBreakpoint";
-export * from "./test-utils";
